Move Iterator implementation to DataValueExt
