+++ cranelift/codegen/src/data_value.rs
@@ -367,33 +367,6 @@ pub fn write_data_value_list(f: &mut Formatter<'_>, list: &[DataValue]) -> fmt::
     }
 }
 
-impl DataValue {
-    /// Returns an iterator over the lanes
-    pub fn iter_lanes(&self, ty: Type) -> DataValueIterator {
-        DataValueIterator::new(self, ty)
-    }
-}
-
-/// Iterator for DataValue's
-pub struct DataValueIterator<'a> {
-    dv: &'a DataValue,
-    ty: Type,
-}
-
-impl<'a> DataValueIterator<'a> {
-    fn new(dv: &'a DataValue, ty: Type) -> Self {
-        Self { dv, ty }
-    }
-}
-
-impl Iterator for DataValueIterator<'_> {
-    type Item = DataValue;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        Some(DataValue::I8(42))
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
@@ -409,15 +382,5 @@ mod test {
             TryInto::<i32>::try_into(DataValue::V128([0; 16])).unwrap_err(),
             DataValueCastFailure::TryInto(types::I8X16, types::I32)
         );
-    }
-
-    #[test]
-    fn test_iterator() {
-        let dv = DataValue::I8(17);
-
-        for d in dv.iter_lanes(types::I8).take(3) {
-            println!("> {}", d);
-            assert_eq!(d, DataValue::I8(42));
-        }
     }
 }
